Impotent coffeescript rewrite.
Because I continue to childishly refuse to write a single line of
"real" javascript.

This is more or less copied line-for-line from the original javascript
with only two notable changes.

The first is to replace a few instances of
`$scope.deck.cards[$scope.currentIndex]` with `$scope.currentCard` as
they should point to the same object. Hopefully this doesn't secretly
ruin everything (it currently doesn't seem to, at least).

The second is to use dot notation for accessing objects (when possible)
rather than bracket notation, and to make this consistent throughout
the script.