Stop trying to load everything as images.
Change the behavior from angular.js hiding/showing text or an image, to
passing the innerHtml for the card div straight from nextCard().

The reason for doing this is primarily the fact that I didn't like it
filling my browser console with load errors on every new card. Other
than that I don't know if it's a better way to handle it.