Stop writing current deck to localStorage
I don't really see what purpose this serves since it's never accessed.
It's maybe there for the future, but I'm dubious of its usefulness.
We're probably better off dumping everything as one json blob under one
key so it only has to be parsed once (on intialization).

Or maybe I'm an idiot and that's a terrible idea.