Store current card index for each deck in localStorage.
Instead of automatically skipping forward when changing decks, change
to the previously used card (or 0, or random, as the case may be).

Also, looking at how localStorage works, it seems that writing
everything we need to store to a single json blob might not be a
completely terrible idea. Everything is stored as strings, and the only
time we should really need to access it is reading (when page is
opened) and writing (just before page is closed).