Don't write long URLs to the answer box or word list.
This loads the full size images in the word list as soon as the deck is
selected. Because the image detection/cropping/resizing occurs when
each card comes up for the first time, there's no good way currently to
do this.

A solution would be to process all images as soon as the deck is
selected, and to this end it might be a good idea to be marking images
(especially on decks with mixed content) in the deck syntax, rather
than determining whether or not a card is an image at runtime.