Fix loading of deck position, write localStorage onunload.
userData can be script-local, since it isn't exposed to angular/the UI
anywhere. It also makes calling userData keys that much shorter (7
characters!).

Drop use of $filter('json') for browser built-in JSON.stringify because
I don't care about browsers that don't have built-in JSON support.

Use window.onunload to write userData to localStorage, which is the
only time it should actually need to be saved, as it stays in memory as
long as the page is open.

Initialize deckIndex so that userdata will be modified before a deck
change occurs, and actually load deckIndex from parsed userData (it
doesn't have its own localStorage key any more).