Stop letting people with quick fingers abuse the answer function.
Because there was a 700ms delay between when the answer was put in the
input box and it repeated the card, pressing <any button> would trigger
a jlearn.check() call, which would evaluate to true because the script
had just written the correct answer to the input, and the input's
maximum length is exactly the length of the answer (meaning whatever
key pressed would not be added to the input, thus not invalidating the
answer).

An alternate solution would be to allow the input box to take
target.length+1 characters, causing anything typed after the solution
is shown to result in a wrong answer, but that's bad because (among
other things) it's prone to causing errors accidentally.